BPAAS-1191: Move logic for chosing the db in metamorph to the cmd package
